refactor(survey-service): move api logic from controllers to respective services
move api logic of some controllers to services for seperation of concerns

gh-1949
