+++ services/survey-service/src/controllers/survey-response.controller.ts
@@ -18,10 +18,8 @@ import {
   RestBindings,
 } from '@loopback/rest';
 import {CONTENT_TYPE, STATUS_CODE} from '@sourceloop/core';
-import {JwtPayload, Secret, verify} from 'jsonwebtoken';
 import {authenticate, STRATEGY} from 'loopback4-authentication';
 import {authorize} from 'loopback4-authorization';
-import {ErrorKeys} from '../enum';
 import {PermissionKey} from '../enum/permission-key.enum';
 import {SurveyResponse, SurveyResponseDto} from '../models';
 import {SurveyRepository} from '../repositories';
@@ -69,29 +67,12 @@ export class SurveyResponseController {
     @param.path.string('surveyId') surveyId: string,
   ): Promise<SurveyResponse> {
     const token = this.request.headers.authorization?.split(' ')[1] as string;
-    const secret = process.env.JWT_SECRET as Secret;
-    const decodedToken = verify(token, secret) as JwtPayload;
-    if (decodedToken.surveyCycleId) {
-      const responders = await this.surveyRepository
-        .surveyResponders(surveyId)
-        .find({
-          where: {
-            surveyCycleId: decodedToken.surveyCycleId,
-          },
-        });
-      const responderEmails: string[] = [];
-      responders?.forEach(responder => {
-        if (responder.email) {
-          responderEmails.push(responder.email);
-        }
-      });
-      const validUser = responderEmails.includes(decodedToken.email);
-      if (!validUser) {
-        throw new HttpErrors.Unauthorized(ErrorKeys.NotAuthorised);
-      }
-    }
 
-    return this.surveyResponseService.createResponse(surveyId, surveyResponse);
+    return this.surveyResponseService.createResponse(
+      token,
+      surveyId,
+      surveyResponse,
+    );
   }
 
   @authenticate(STRATEGY.BEARER, {
