+++ services/survey-service/src/services/index.ts
@@ -1,4 +1,5 @@
 export * from './question-helper.service';
+export * from './survey-question.service';
 export * from './question-duplicate-helper.service';
 export * from './question-template.service';
 export * from './question-option.service';
